New config format!
The old INI format was not sufficient for the complexity required by the
bot. The servers setting in core, for instance, was always way too long
for multi-network bots, and was only going to get worse as more features
were added.

The new format uses nested config blocks. Servers are now defined in a
servers block rather than in a single, long string in core.