Use canonized channel names where appropriate in Channels and the channels script. This should resolve the bug that causes Terminus-Bot to constantly issue JOINs for channels that it thinks it should be in but isn't.
Also added a thing to the channels script that will canonize channels
that are already in the database. The old behavior was to just
String#downcase everything, but this was incorrect, and wasn't even
implemented everywhere.

To update a running bot, run:
  reload channels
  raw JOIN 0
  joinchans

(This seems to leave non-canonized channel names in the list. The
easiest fix is to reconnect the bot.)