ircmacros: add send_command function
This command provides further abstraction between the application code
and the IRC connection. The command, parameters, and trailing text
argument are passed as function parameters.

The rationale for this is to allow for prioritization based on message
type (PONG should be send immedaitely, PRIVMSG and NOTICE should be
high-priority, and things like WHO can be sent with low priority). That
functionality isn't yet present, but will be soon.